상세보기,수정,삭제합치기완료

+++ src/main/java/com/example/recipe2022/service/ReplyService.java
@@ -42,9 +42,19 @@ public class ReplyService {
     }
 
     //삭제
-    public ResponseEntity<?> deleteReply(ReplyDto.replyDelete replyDeleteDto) {
+    public ResponseEntity<?> deleteReply(Authentication authentication, ReplyDto.replyDelete replyDeleteDto) {
         if(!replyRepository.existsById(replyDeleteDto.getReplySeq())){
             return response.fail("댓글을 찾을 수 없습니다. ", HttpStatus.BAD_REQUEST);
+        }
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+        String email = userDetails.getUsername();
+        if (userRepository.findByEmail(email).orElse(null) == null) {
+            return response.fail("해당하는 유저가 존재하지 않습니다.", HttpStatus.BAD_REQUEST);
+        }
+        Users users = userRepository.findByEmail(email).orElseThrow();
+        Reply currentReply = replyRepository.findByReplySeq(replyDeleteDto.getReplySeq());
+        if (!users.getReplies().contains(currentReply)) {
+            return response.fail("해당 권한이 없습니다.", HttpStatus.BAD_REQUEST);
         }
         replyRepository.deleteById(replyDeleteDto.getReplySeq());
         return response.success("성공적으로 삭제되었습니다.");
